reafactor: split utils to dedicated helper files
