+++ apps/evm-bridge/tests/helpers/browser.ts
@@ -66,3 +66,23 @@ export const test = base.extend<ExtensionFixtures>({
 });
 
 export const expect = test.expect;
+
+export async function closeBrowserTabsExceptLast(browserContext: BrowserContext) {
+    const pages = browserContext.pages();
+    if (pages.length > 1) {
+        for (let i = 0; i < pages.length - 1; i++) {
+            await pages[i].close();
+        }
+    }
+}
+
+export async function getExtensionUrl(browserContext: BrowserContext): Promise<string> {
+    let [background] = browserContext.serviceWorkers();
+
+    if (!background) {
+        background = await browserContext.waitForEvent('serviceworker', { timeout: 30000 });
+    }
+
+    const extensionId = background.url().split('/')[2];
+    return `chrome-extension://${extensionId}/ui.html`;
+}
