+++ apps/evm-bridge/tests/utils/constants.ts
@@ -1,0 +1,11 @@
+export const WALLET_CUSTOMRPC_PLACEHOLDER = 'http://localhost:3000/';
+
+export const THREE_MINUTES = 180_000;
+
+// Mnemonic for the native token tests
+export const MNEMONIC =
+    'mom program scrap easily doctor seed slender secret mad flat foam hospital cherry seek river you obscure column blood reflect arch pencil cat burst';
+export const TOOL_COIN_TYPE =
+    '0xe1e88f4962b3ea96cfad19aee42f666b04bbce4dc4327c3cd63f1b8ff16e13b2::tool_coin::TOOL_COIN';
+export const TOOL_COIN_OBJECT_ID =
+    '0xf7662ffd9cb079d8e75ab4805ba78fdb0e0fb78cf49aa0fa01ecb7ebdf15d04e';
